Add configuration management and audio processing modules with initial setup; Added redundancy option with balancing
