+++ server/public_ip_or_local_network/configs/selected_client.json
@@ -1,0 +1,3 @@
+{
+    "selected_client_ip": "34.145.14.145"
+}
