+++ server/public_ip_or_local_network/configs/server_config.json
@@ -2,5 +2,5 @@
     "redundancy_level": 1,
     "auto_select_first": true,
     "use_ipv6": false,
-    "use_ssl": false
+    "use_ssl": true
 }
